211--update webpack configs and add lazy wrappers for all templates

+++ packages/editor/src/main/models/projects/projects.ts
@@ -973,18 +973,35 @@ export const open = (ev: rq.RequestEvent, project: ProjectMeta) => {
 
       const pathName = fs.joinPath(fs.getDirname(project.filename), 'assets');
 
-      fs.copy(pathName, fs.APP_PATHS.uploads).then((copyRes) => {
-        if (copyRes.error) {
-          resolve(copyRes);
+      fs.fileExists(pathName).then((existsRes) => {
+        if (existsRes.error) {
+          resolve(existsRes);
           return;
         }
 
-        resolve({
-          error: false,
-          data: {
-            project: JSON.parse(res.data.contents),
-          },
-        });
+        if (!existsRes.data.exists) {
+          resolve({
+            error: false,
+            data: {
+              project: JSON.parse(res.data.contents),
+            },
+          });
+          return;
+        } else {
+          fs.copy(pathName, fs.APP_PATHS.uploads).then((copyRes) => {
+            if (copyRes.error) {
+              resolve(copyRes);
+              return;
+            }
+
+            resolve({
+              error: false,
+              data: {
+                project: JSON.parse(res.data.contents),
+              },
+            });
+          });
+        }
       });
     });
   });
